revert annot and src span changes
